Improve the proof of ObjectMap and ObjectArray, add auxiliary methods, etc.
