+++ src/components/ndn-lib/src/coll/mod.rs
@@ -1,0 +1,3 @@
+mod def;
+
+pub use def::*;
