+++ src/components/ndn-lib/src/object_map/mod.rs
@@ -3,11 +3,12 @@ mod storage;
 mod memory_storage;
 mod file;
 mod storage_factory;
+mod proof;
 
 pub use object_map::*;
 pub use storage::*;
 pub use memory_storage::*;
 pub use storage_factory::*;
-
+pub use proof::*;
 #[cfg(test)]
 mod test;
